Move restriction e2e tests to the addon API submission flow
Following some feedback the end to end test were driving legacy signing
APIs, which we eventually want to move away from. They now live in
TestVersionViewCreate and submit through the v5 addon api.

The exact same coverage as before, just moved locations.
